paraview: Downgrade to version 5.6.3
Paraview 5.7.0+ has a bug relating to polygon and line offsetting which
means that when viewing a "Surface With Edges" representation at high
zoom excessive amounts of edges that should not be visible are shown.
This makes inspection of a typical mesh almost impossible.

See issues 19723 and 19437 on ParaView's gitlab.

Downgrading to version 5.6.3 until this issue is resolved.