equationOfState: Separated entropy contributions into Sp and Sv
The entropy contribution from the equation of state corrects the integral of the
heat capacity divided by temperature for changes in pressure in the case of Cp
named Sp or changes in volume in the case of Cv, named Sv.  This for enthalpy
based thermodynamics Sp is needed and for internal energy Sv is needed.