calculatedFvPatchField: Changed fixesValue to return false
While calculatedFvPatchField would be considered as a BC which fixes the value
during solution it cannot be used for that purpose and under all other
conditions it does not fix the boundary value as the '=' operator changes the
value.  However calculatedFvPatchField is used for the reference phase in
multiphase systems and if it is considered to fix the boundary value it interferes
with the operation of the MULES limiter and hence it is far preferable if it
does not fix the value.