multiphaseEulerFoam: Improved limitSum for cases with stationary phases
limitSum operates on the sum positive and negative flux corrections as it did
originally to guarantee that the phase fractions sum to 1 but now on the scaled
moving sub-set of the phases so that it handles the presence of stationary
phases in a consistent manner.

Additionally limitSum is now applied to two-phase systems even when only one of
the phases is solved for to ensure the solution is the same irrespective of
which phase fraction is solved or if both are solved.

compressibleMultiphaseInterFoam and multiphaseInterFoam have been updated to use
the same form of limitSum as multiphaseInterFoam but this does not change their
behaviour, it is to reduce code duplication.