fvPatch: Removed unnecessary default arguments in lookupPatchField
These were marked as being necessary to resolve a bug in Gcc, but no
tested version of Gcc failed to compile with these removed. It is
assumed that the Gcc bug is in a version older than those supported.

Also added "db()" method to fvPatch in order to remove inclusion of
fvPatchFvMeshTemplates.C in fvMesh.H. The resulting inclusion strategy
is significantly simpler as a result.