rhoTabulated, hTabulatedThermo, tabulatedTransport: New (p, T) tabulated thermophysical functions
This is a prototype implementation of (p, T) tabulated density, enthalpy,
viscosity and thermal conductivity using a uniform table in pressure and
temperature for fast lookup and interpolation.  The standard Newton method is
used for h->T inversion which could be specifically optimised for this kind of
table in the future.