solutionControl: Execute function objects after residual check
Function objects that write need to execute after the residual checks so
that the associated modification of the time state correctly makes them
write on the final iteration.

This reverts commit c6089ae3a65f608a60d665c2baa45158ce954789, meaning
that residual controls can no longer apply to fields solved for as part
of function object executions.

Creating a system which permits solutions done within function objects
to be subject to residual control, *and* which correctly causes function
objects to write on final iterations, would be significantly more
complex.