topoSet: searchableSurfaceToFaceZone: Improved robustness
This set now produces a contiguous face zone even when cell centres fall
exactly on the plane of the searchable surface.

The set now calculates the intersections between the
cell-centre-to-cell-centre vectors for every face, and uses those to to
categorise the adjacent cells either side of the set as either above or
below the surface. Then, faces which border both an above and a below
cell are included. This guarantees a contiguous result.

Previously faces for which an intersection was found were included
directly, which resulted in a non-contiguous result when the
intersections became ambiguous; i.e., when the surface passed very close
to the cell centres.