snappyHexMesh: Fixed syncronisation of position averages
Sums of positions cannot be syncronised directly, as they are neither a
true position or a displacement, so that cannot be transformed across
coupled interfaces. The difference to the point/edge/face should be
summed instead, and syncronised as a displacement. The point/edge/face
location can then be added on again after syncronisation.

This resolves bug report https://bugs.openfoam.org/view.php?id=3602