fvOptions: Enabled lookup and evaluation of individual options
This change allows fvOptions to access and evaluate each other. This
means an option can be written that applies a source value proportional
to that in another option.

For example, you might have an fvOption specifying a mass source term in
the rhoEqn. You could now write a corresponding momentum source fvOption
which takes the velocity associated with that mass source as an input.
It would look up the mass source option to get the mass flow rate and
multiply it by the user supplied velocity to get the momentum source.
This prevents the user having to convert from the primitive variable
(e.g., velocity) to the conserved variable (e.g., momentum) when
specifying source terms.

Lookup is provided by means of a new [](const word&) operator on the
fvOptionList, and evaluation is achieved by adding ()(...) operators to
the fvOptions similar to those already available in the fvOptionList.
Example usage is shown below.

    const fv::optionList& fvOptions(fv::options::New(mesh_));
    const fv::option& fvOption = fvOptions[sourceName_]; // <-- Lookup
    tmp<volScalarField> source(fvOption(field) & field); // <-- Evaluation