porosityModel: Rationalised and simplified the interface
interRegionExplicitPorositySource: Removed the unnecessary and inappropriate
bidirectional mapping of data between the regions sharing the porous media.  Now
the drag is calculated on the same mesh as the flow and filtered according to
the fraction of the cells containing porosity.

To complete this work a small change will be made to the mesh-mapping functions.